pyproject.toml
[project]
name = "beangrid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "lark>=1.2.2",
    "pydantic>=2.11.7",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
